más cambiosh
ouya

+++ gerente/gerente/src/main/java/com/example/gerente/model/Factura.java
@@ -27,7 +27,7 @@ public class Factura {
     @Min(value = 0)
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_factura;
+    private int id_factura;
     @Column(name = "fecha",nullable = false)
     private Date fecha;
     @Column(name = "total",nullable = false)
